scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "python-dotenv>=1.1.0",
    "typer>=0.15.3",
]

[dependency-groups]
dev = [
    "asyncpg-stubs>=0.30.1",
    "mypy>=1.15.0",
    "ruff>=0.11.7",
]
